--- SeSAC Shopping/View/TableViewCell/MainTableViewCell.xib
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="22505" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="22504"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="MainTableViewCell" rowHeight="53" id="KGk-i7-Jjw" customClass="MainTableViewCell" customModule="Recap_Assignment" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="384" height="53"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="384" height="53"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="ㄴㅇ" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ic4-tC-Xa9">
-                        <rect key="frame" x="55" y="5" width="250" height="43"/>
-                        <constraints>
-                            <constraint firstAttribute="width" constant="250" id="dzl-OM-whN"/>
-                        </constraints>
-                        <fontDescription key="fontDescription" type="system" pointSize="13"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="YGb-5X-aMh">
-                        <rect key="frame" x="15" y="5" width="25" height="43"/>
-                        <constraints>
-                            <constraint firstAttribute="width" constant="25" id="OLb-qU-EAd"/>
-                        </constraints>
-                    </imageView>
-                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="qsc-uA-t4J">
-                        <rect key="frame" x="325" y="10" width="49" height="33"/>
-                        <state key="normal" title="Button"/>
-                        <buttonConfiguration key="configuration" style="plain" title="Button"/>
-                    </button>
-                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="cuc-0R-jae">
-                        <rect key="frame" x="55" y="5" width="250" height="43"/>
-                        <state key="normal" title="Button"/>
-                        <buttonConfiguration key="configuration" style="plain" title="Button"/>
-                    </button>
-                </subviews>
-                <constraints>
-                    <constraint firstItem="qsc-uA-t4J" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="10" id="123-Jd-vbw"/>
-                    <constraint firstItem="YGb-5X-aMh" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="15" id="2c5-Lx-7kr"/>
-                    <constraint firstItem="qsc-uA-t4J" firstAttribute="leading" secondItem="ic4-tC-Xa9" secondAttribute="trailing" constant="20" id="3OS-VG-5bt"/>
-                    <constraint firstAttribute="bottom" secondItem="YGb-5X-aMh" secondAttribute="bottom" constant="5" id="3rn-Mb-Z0T"/>
-                    <constraint firstItem="cuc-0R-jae" firstAttribute="bottom" secondItem="ic4-tC-Xa9" secondAttribute="bottom" id="9pA-0a-gEs"/>
-                    <constraint firstAttribute="trailing" secondItem="qsc-uA-t4J" secondAttribute="trailing" constant="10" id="R4b-RC-MVg"/>
-                    <constraint firstAttribute="bottom" secondItem="ic4-tC-Xa9" secondAttribute="bottom" constant="5" id="RId-80-5q4"/>
-                    <constraint firstItem="cuc-0R-jae" firstAttribute="leading" secondItem="ic4-tC-Xa9" secondAttribute="leading" id="UyH-lY-vob"/>
-                    <constraint firstItem="cuc-0R-jae" firstAttribute="trailing" secondItem="ic4-tC-Xa9" secondAttribute="trailing" id="WDw-8s-RQu"/>
-                    <constraint firstAttribute="bottom" secondItem="qsc-uA-t4J" secondAttribute="bottom" constant="10" id="bWl-CF-CGo"/>
-                    <constraint firstItem="cuc-0R-jae" firstAttribute="top" secondItem="ic4-tC-Xa9" secondAttribute="top" id="k4K-WR-S0t"/>
-                    <constraint firstItem="ic4-tC-Xa9" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="5" id="s2R-KC-bQz"/>
-                    <constraint firstItem="ic4-tC-Xa9" firstAttribute="leading" secondItem="YGb-5X-aMh" secondAttribute="trailing" constant="15" id="sMn-QA-puD"/>
-                    <constraint firstItem="YGb-5X-aMh" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="5" id="w9B-bM-qKc"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="mainCellButton" destination="qsc-uA-t4J" id="cCg-8O-EIm"/>
-                <outlet property="mainCellClickedButton" destination="cuc-0R-jae" id="gVq-KW-XTZ"/>
-                <outlet property="mainCellImageView" destination="YGb-5X-aMh" id="d7Z-dc-rUv"/>
-                <outlet property="mainCellLabel" destination="ic4-tC-Xa9" id="W4c-iG-1gp"/>
-            </connections>
-            <point key="canvasLocation" x="175.57251908396947" y="-8.8028169014084519"/>
-        </tableViewCell>
-    </objects>
-</document>
